backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.1
SQLAlchemy==2.0.35
alembic==1.13.2
jinja2==3.1.4
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
